Remove run-scoring, wrapper, cost-proxy, and smoke-harness surfaces
These are leaf features of the legacy guarded-run lane with no consumer in the
current product: the advisory judge/value scoring, the agentacct-claude /
agentacct-codex process wrappers, the opt-in cost proxy, and the optional
real-agent smoke harnesses. Removed together with their CLI sub-apps, MCP
tools (prepare_judge/compute_value), REST routes (/runs/{id}/judge/prepare,
/runs/{id}/value/compute), the provider/API-proxy usage truth-table row, and
their tests. outcome.py is reduced to the machine-check helpers still used by
record_machine_check.

The run/report/record_machine_check core is untouched: the run/report/runs/
pause/resume/kill/scan/demo commands, runner.py, reports.py, the /runs read
routes, service.list_runs/get_report, the MCP list_runs/get_report/
record_machine_check tools, and the process-wrapper truth-table row (which
documents `agentacct run`) all stay. A separate change removes the guarded-run
core, which requires changing the record_machine_check tool contract and the
receipt ledger feed.
