+++ tests/test_usage_truth.py
@@ -30,7 +30,6 @@ def test_usage_truth_table_has_expected_integrations() -> None:
     assert "OpenLIT / OTLP trace ingestion" in rows
     assert "Paperclip snapshot connector" in rows
     assert "Entire Git checkpoint connector" in rows
-    assert "provider/API proxy" in rows
     assert "sentinel-owned process wrapper" in rows
 
     assert rows["codex local usage import"]["usage_confidence"] == "client_reported"
@@ -47,7 +46,6 @@ def test_usage_truth_table_has_expected_integrations() -> None:
     assert rows["agent MCP workflow events"]["usage_confidence"] == "unknown"
     assert rows["native coding-agent hook capture"]["usage_confidence"] == "unknown"
     assert "not proof of billing" in " ".join(rows["OpenLIT / OTLP trace ingestion"]["limitations"])
-    assert "hard dollar enforcement" in rows["provider/API proxy"]["hard_budget_basis"]
     assert "optional fallback" in " ".join(rows["sentinel-owned process wrapper"]["limitations"])
 
 
@@ -58,7 +56,6 @@ def test_usage_truth_table_cli_json() -> None:
     payload = json.loads(result.output)
     integrations = payload["integrations"]
     assert len(integrations) >= 6
-    assert any(row["integration"] == "provider/API proxy" for row in integrations)
     assert "provider_billed" in json.dumps(payload)
     assert "subscription billing" in json.dumps(payload)
 
@@ -71,7 +68,6 @@ def test_usage_truth_table_cli_human_output() -> None:
     assert "codex local usage import" in result.output
     assert "Freshness" in result.output
     assert "agent MCP workflow events" in result.output
-    assert "provider/API proxy" in result.output
 
 
 def test_usage_truth_table_is_documented() -> None:
